requirements.txt
selenium
python-dotenv
csv